[carnelian] Create a canvas display mode
Created a mode for Carnelian apps that can do all of their rendering
with Canvas to use and be able to run under both Scenic and the software
frame buffer provided by the Display Manager. This view mode is
specified by the ViewMode enum that is returned by the get_mode() method
of the AppAssistant trait. This design means that an application cannot
have both Scenic and Canvas views at the same time. For now this seems
like a reasonable simplification.

Modified the drawing example to be the first of these dual-mode
applications.

Modify the PixelSink trait to be just a use of SharedBuffer in actual
rendering but maintain PixelSink for testing purposes. Switched to the use
of Mapping everywhere to remove low-level Zircon calls.

Modified ViewAssistantContext to provide optional resources for for
both modes.

Created an AppStrategy trait to encapsulate the data and behaviors
for apps running in frame buffer mode or scenic.

Create a ViewStrategy trait to encapsulate the data and behaviors for
views running in Scenic, Scenic Canvas and frame buffer canvas modes.

Added class/display-controller and class/input to the dev section of
the service manifest uses for examples to support the drawing sample
under frame buffer mode.

Tested by running all Carnelian dependencies under Scenic/tiles or
frame buffer.

Also tested via the included integration test.

Change-Id: Id7cf947edf69fbad2ff85f0873e0c593a93c5a33