[scenic] Inject Escher into GfxSystem
A partial implementation of SCN-1491, this CL also fixes SCN-568 and (hopefully) SCN-746.

It co-ops the System class to represent dependencies (instead of adding new logic for initialization) as this logic may change drastically with the next dependency injection (namely, Engine), and is likely to be refactored entirely (see SCN-1504).

tested=gfx_pixeltests, scenic_unittests

Change-Id: I0d96eaa7d85bdaf5cc07259f6b79f09f3c644d32