[debugger] Make debug agent persists upon disconnects.
Before the debug agent would be bound to a connection, so the moment the
connection got lost, the agent would be destroyed and all the
attachments/breakpoints would be lost, somewhat defeating the purpose of
reconnecting.

With this change, the agent works independent of a connection, so can
connect/disconnect multiple times without loosing state.

IMPORTANT: While the agent doesn't loose state, this setup will actually
           "leak" as the newly connected agent won't know about the current
           state of the debug agent (it assumes a "clean" debug agent).

           While this is somewhat a regression, it's not that bad as
           the default setup is to quit on exit so the agent will exit
           anyway upon disconnection.

           That change will come in a subsequent CL(s).

TEST=Manual with debug mode activated agent and using `quit-on-exit`
     false config, which enables an agent to remain upon disconnection.

Change-Id: If78776a9a215e42613446ec773b5ccfd24c7cfee