[scenic] change frame scheduler offset from 1->3ms
With this change, we expect frame rates to at least stay constant, but
most likely increase.

FPS should become more stable as we allow for more variance in the time
it takes to produce a single frame.

However, the trade off is the lowest latency we can achieve increases
slightly, so it is possible that input latency benchmarks may incease
slightly as well, but by no more than 2ms on average.

Test: More consistent FPS across various hardware, including
astro, pixelbook, and AEMU. Unit tests.

Change-Id: I62d0cd95746cd0aeb6dc58ec5c6dea9272c8f659