[banjo][syzkaller] Support for syzkaller structs, unions and flags
This CL adds the capability to delcare and use structs, unions and enums
as method arguments. Declarations are translated to syzkaller-supported
format[0] for structs, unions and enums(called flags in syzkaller).

For now, syscall declarations remain the same and arguments are linked
to their respective enums/unions using attributes('valueset'/'typeset').
Common argument processing code for methods, structs and unions was
refactored, and it uses class memebers for state keeping purposes.

[0]: https://github.com/google/syzkaller/blob/master/docs/syscall_descriptions_syntax.md

Test: fx set core.x64 --with //bundles:tests
      fx run-host-tests banjo_unittests

SEC-348 #done

Change-Id: I7b5e6a77dbcf10d8d2ab1947ee456a9fdfcf218b