[debugger] Add std::string pretty-printer.
This is a hardcoded pretty printer for our libc++'s std::string
implementation. It is designed to work around the problem where we have
no symbols for std::string, and also format it in a nice way.

Eventually this should be rewritten in a more extensible pretty-printing
system that can be versioned with our library rather than with the
debugger binary.

Change-Id: I3d6095b57790872e548fc876ddc09bdce3e6c9e7