[kernel] Manipulate WaitStateObserver::wakeup_reasons_ with atomics
The WaitStateObserver's wakeup_reasons_ member type is mutated from
multiple threads through the StateObserver callbacks. The mutating
callbacks are serialized through one dispatcher lock, so only one will
occur at a time.  When one of these mutations trigger a wakeup, an event
is signaled which will lead to the End() call executing on a different
thread from mutators.  It's important that the End() call observe a
value that was set and that was sufficient to trigger a wakeup.  The
specific value observed by End() doesn't matter in the case of repeated
mutations.

This upgrades the wakeup_reasons_ type from a raw primitive
(zx_signals_t) to an ktl::atomic<> and replaces the stores and loads
with atomic operations using the acquire / release memory order.

Also adds a defensive assert to the object_wait_one path to verify that
the signal set returned by the WaitStateObserver is consistent with the
return code.

ZX-4670 #done

Change-Id: I1537f01055ba0dd8c7e4af8d0873378454324254