[codec] Remove double lock in failure scenarios
CodecImpl::FixupBufferCollectionConstraints is always called with the
lock already held. Typically, this is fine, but in failure scenarios, it
was calling CodecImpl::Fail rather than CodecImpl::FailLocked, which
meant that we would try to acquire the lock again. This only manifested
when the CodecAdapter provided invalid BufferCollectionConstraints. To
clarify the requirements, the Locked suffix has been added to
FixupBufferCollectionConstraints and all Fail calls have been updated to
FailLocked.

Change-Id: I822f12d968a15f35e85d7d6a4d9d989815c06ac3