[clang-tidy] Allow structs with public members.
The check "misc-non-private-member-variables-in-classes" currently
disallows struct and classes which public fields if they have at least
one method on them. For example, the following is disallowed:

  struct Foo {
    int a;
    int b;
    int c;

    string ToString() const;
  };

This is perhaps more drastic than intended. Switch the option
"misc-non-private-member-variables-in-classes" to only disallow public
fields on structs and classes if at least one field is private.

Change-Id: Ifbb98e95291e8e10084459249a500b6d79503383