[lib][affine] Reduce dependencies and stop auto-reducing ratios.
Reduce the number of dependencies that libaffine brings in right now
so that it may be used in the vDSO image for converting ticks to clock
monotonic timestamps.  In particular..

++ Replace usage of ASSERT/DEBUG_ASSERT with internal version which do
   not depend on printf, and do no depend on abort.  This is to avoid
   bringing in a non-header libc dependency.
++ Remove the dependency on third_party/safemath.  We were using very
   little of it.  Instead, use the compiler intrinsics for detecting
   overflow, and create clamped addition/subtraction routines which we
   can us instead.
++ Make sure to flag this library with the magic "visibility_hidden"
   configuration.  This helps to prevent symbols from this library
   being unintentionally exported from a DSO which links against this
   static library.

In addition to this, stop automatically reducing 32 bit ratios.  There
really is not much benefit in doing so.  Either the
numerator/denominators fit in 32 bit storage, or they don't.  During
scaling of a 64 bit integer, the work remains the same.

Also, remove the == and != operators for Transform and Ratio.  Simple,
low cost tests for equality rely on an expectation of ratios always
being reduced, which they no longer are.  Instead, provide an in-place
Reduce method for ratios allowing users to perform the reduction
themselves if they _really_ want to test for equality.

Update tests for the behavior changes, and introduce new tests for the
in-place reduction method, and for the clamped add/sub routines..

Finally, put clang-format on/off directives around the test vector
tables in preparation for the great reformatting.

Tests: updated unit-tests.
Change-Id: I2d2c254b14daa9b2e12dc0aac5b80fdcf56804de