[arm64][kernel] Clear exclusive monitor in context switch
The exclusive monitor is logically part of a thread's context.  Clear
it to ensure the new thread does not complete the store half of a
load-exclusive/store-exclusive started by the old thread.

While investigating ZX-4595, travisg@ and I noticed that we don't
clear the monitor in arm64_context_switch.  While this does not appear
to be the cause of ZX-4595, and is unlikely to cause problems, we
should clear it for correctness.

Imagine there is one CPU and two threads, A and B.  Both threads are
atomically incrementing the value pointed to by x9 (think
value.fetch_add(1, std::memory_order_relaxed)).

A:
    ldxr x1, [x9]
    add  x10, x10, #0x1
    stxr x11, x10, [x9]
    cbnz x11, A
    b    yield

B:
    ldxr x1, [x9]
    add  x10, x10, #0x1
    b    yield
    stxr x11, x10, [x9]
    cbnz x11, A

Imagine that the yield function may perform a context switch.  Say A
has performed a successful store-exclusive and is now about the call
yield.  The CPU's monitor may or may not remain in the Exclusive
state.  Let's say it does.  We now call yield and context switch to B.

When B resumes execution the next instruction executed is the stxr.
Remember, A's store to the same address has just succeeded.  Given
that the monitor is still Exclusive, when we execute B's stxr, it will
succeed even though A has perform an intervening operation (!).

The scenario above is fairly contrived and unlikely to happen in
practice.  It's not common for a thread to voluntarily yield in the
middle of a load-exclusive/store-exclusive operation.  Furthermore, an
eret will clear the monitor and so if the old thread had been
preempted, then the eret performed before resuming it will clear the
monitor.

Test: booted vim2 and qemu+kvm on Cavium Thunder X2
Bug: ZX-4595 #comment clrex
Change-Id: I5a7f0fb0157826607c4212fd847e36787edb606b