[kernel][pmm] Support asynchronous page allocation
This change adds a flag that causes synchronous allocations to fail
during out of memory situations and it adds apis to create and manage
asynchronous allocation requests. It also adds an api to register a low
memory callback with the pmm, so that the system can reclaim memory to
prevent falling into an out of memory state or to recovery from such an
out of memory state.

Asynchronous page allocations are fulfilled by a dedicated thread in
PmmNode. It may be possible to fulfill some delayed requests
synchronously with pages being freed, but doing so would require
significant work to prevent lock order violations. Delayed requests
are fulfilled in strictly FIFO order. However, they are not given
special priority over incoming synchronous requests (e.g. by directly
passing pages to the allocation callbacks), as it is not expected that
the system remain in an out of memory state for an extended period of
time.

This change also deletes the oom library and replaces its usage with the
newly introduced pmm low memory callbacks. This change replaces the old
'k oom' commands with 'k pmm mem_avail_state' commands.

Test: k ut pmm
Test: Trigger ooms on astro by consuming memory with dd, as well as by
manually invoking the new 'k pmm mem_avail_state' command.

Change-Id: I89abaf150fbf1971d7043a6cb3355de77aad2b8f