[bt][tools] Add disconnect command to bt-cli
Peers can now be disconnected from the command line through usage of the
disconnect command:

  disconnect PEER

where PEER is either an identifier, or the address of a peer known to
us allowing us to look up the identifier.

testability: Added unit tests to assert that entered command lines
correctly parse to the disconnect command and the expected PEER id that
will be sent to the control service.

The control service and host driver implementations have their own
integration tests, added in CLs 290478 and 290048 respectively.

Change-Id: Ie80b7d0eee98a28e4556b2350252b718805fa0a0