[a11y] Defines a base class for a Screen Reader.
This change defines a base Screen Reader class.  Methods and internal logic are
not implemented and will be added in future changes.

It is submitted as it is for the sake of making easier the process of linking
some accessibility code being developed, where the Screen Reader is a central
component.

MI4-2536 #comment

Test: installed image in device.

Change-Id: I0f7974df42161c20f80c18a91ef102e369bef66b