[zircon][dev][goldfish] Add Pipe protocol with Call method
The Call method provides one important optimization; the
ability to write and readback a result without having
to make 4 context switches. This is a very common pattern
required by the goldfish vulkan ICD.

This optimization provides a 30% performance improvement
for vulkan API calls that fall into this pattern.

In order to implement this optimization, pipe processing
has been moved to a seperate thread. This is required
to not block the main message loop of the devhost when
processing Call methods.

Thread changes are covered by existing tests and
GoldfishPipeTest has been updated to verify the behavior
of the new Call method.

Note: goldfish vulkan ICD is the only user of this API
today. Soft transition can be avoided by ensuring that
an updated ICD is rolled in with this change.

Test: goldfish-test
Test: goldfish-benchmark

Change-Id: I5b239014a6b1e53e9b6a437956a7b2b1bbb92f02