[codec][drm] Add decryptor support to codec_impl
In order to create a decrypting StreamProcessor, we need to teach
CodecImpl (which is essentially StreamProcessorImpl) to accept
fuchsia::media::drm::DecryptorParams as possible input parameters. This
implies that CodecImpl will now be capable of acting as either an
Encoder, Decoder or Decryptor, but cannot do so concurrently. Depending
on which type of params are provided at construction, the appropriate
behavior will be selected.

Previously, the CodecImpl was capable of being either an Encoder or
Decoder and had a different constructor for each. This CL also collapses
the constructors into one by using a variant for the parameter type. A
single constructor will now accept either the
fuchsia::mediacodec::CreateEncoder_Params, CreateDecoder_Params or
fuchsia::media::drm::DecryptorParams.

Change-Id: I46d3c7ff363a2bacce94c5f334d416b801c38c75