[dev] Start ethernet device with expected settings.
Make sure a new device is started with known settings.
For now, it's only that the interface is not in promiscuous mode.

BUG: NET-2439
TEST: Made sure interface is not in promiscuous mode by sending
packets to an SMAC different from the interface's MAC and verifying those
packets are not received.

Note that this currently does not work on Anker adapter, as that driver
does not currently support changing its configuration.
Created http://fxb/ZX-4655 to track the driver work needed.

Change-Id: Ibf222df0fe2e06af5bfc9cee2200f7df9e610bb3