[fake-bti] Return non-zero physical addresses
The io-buffer library treats zero as an invalid physical address, which
makes it incompatible with fake BTI. This change makes the fake BTI
return "validish" physical addresses from zx_bti_pin, i.e. an arbitrary
address that is page aligned and not zero.

This arbitrary value is exposed as FAKE_BTI_PHYS_ADDR in case any tests
want to check that a driver correctly passes a physical address to a
device by mocking out register writes (for example).

Note that previously a decision was made to treat zero as an invalid
physical address, even though technically it might be valid, because it
is common for drivers to zero-allocate io buffers and treating zero as
invalid catches the unitialized case.

Change-Id: I73dd4cacb9cae175f7038b40226a5ff2cc986526