[scenic] Convert AddPart to AddChild
Anywhere within scenic/escher where we call
node->AddPart() has been converted to
node->AddChild(). This is the first step in
an ongoing effort to deprecate node parts.

No API changes have been made yet, meaning
that as of this CL "AddPart()" is still callable.

Testing - This is a pure refactor, which according to
the testability rubric does not require a unit test.

Change-Id: I82c9c5f5948691babc5ea97d2dcebd1104d32d38