[bt][gap-bredr] Drive pairing state transitions using event handlers
Add skeleton event handlers to PairingState for the HCI events it will
respond to. The handlers have no side effects but signaling the wrong
HCI event on a state causes an externally-observable error.

Test that the state transitions occur as expected and errors out when
events are not received in the order expected.

BT-374

Test: In bt-host-unittests
GAP_PairingStateTest/HandlesEvent.*

Change-Id: I2125ef74a8d065a904c95486af9ea4ecf46927d2