[bootsvc] Update bootsvc.next default to start component_manager
Standalone Zircon builds have been deprecated as of BLD-382 completion
on 20190703, so we can remove the bootsvc.next command line override in
bringup.gni to start component_manager and make it the default in
bootsvc itself. The only practical impact of this change is that Zedboot
images will start using component_manager.

Bug: CF-659 #done
Change-Id: I881501ee02aff8e27cbc5e6bff3fac80ff4859ad