[virtmagma] add wayland importer
This change adds a new VirtioWaylandImporter interface through which a
VirtioWayland device can import a VMO for use as a VFD. It also exposes
this as an optional argument to VirtioMagma. If present, all buffers
passed to magma_export will be subsequently imported into the wayland
device. This is currently only exercised when using the vulkan wayland
swapchain extension.

This change also introduces two new device tests - one for magma export,
and another for wayland import. A subsequent change will include the e2e
swapchain test.

Test: ran wayland_swapchain_test in biscotti_guest
      ran 'fx run-test vmm_tests'
Change-Id: I7655caf774aa098f75606ce54242fb0eda30b24a