[zircon] Prepare for the great reformatting.
Add a bunch of clang-format on/off comments for table centric code
that I would like to preserve the formatting for.  For the most part,
this tends to be stuff like tables of register definitions, and test
vectors for test code.  Note that in a few places where it matters,
the indentation inside of the format off section has been aligned to
2-space indentation so that things will be consistent once the grand
reformatting is complete.

Also note that, in a couple of limited cases, undo some places where
this code had already been auto-formatted.  In these cases, revert to
the pre-formatted form, add the clang-format directives, then
re-format using fx format-code.

Change-Id: Ic8fb1a0d5ef72dbd991aa482ecad612c7cb1bba1