[zx][clock] Introduce lib/zx/clock.h
Prepare to move zx::clock out of time.h and into it's own clock.h.
Right now, zx/object.h depends on the definition of zx::time (defines
in time.h), but zx::clock is also defined in time.h.  In order for
zx::clock to become an object based on zx::object<>, this dependency
cycle needs to be broken.

Start by adding lib/zx/clock.h to the system.  For now, simply have it
inlclude lib/zx/time.h.  The next step will be to update the current
users of zx::clock so that they properly include clock.h.

Change-Id: I82e46e4692c168f78faa8aa65c2ed25040149d35