[kernel][fbl] GPArena library
Introduces a Growable Persistant Arena (GPArena) datastructure. This
is similar to the existing fbl::Arena except that the GPArena provides
a guarantee that portions of the object will be preserved between a
Free+Alloc.

A consequence of this guarantee is that memory used by the arena will
never be reclaimed and will monotonically grow, removing the need for
any code to handle decommit and allowing for a trivial lock-free free
list of objects that can reuse the object memory that is not marked for
preservation.

Change-Id: I18f0c9d944f28a378aaeebbb8053072a658d8fef