[kernel] Warn if cpus we try to start don't start.
On ARM platforms the cpus are started based on information provided by
the bootloader and that information could be wrong.

If we try to start CPUs and they don't come up we will issue a warning
in the logs as the performance of the system will be far from what is
expected.

We do this by giving the cpus several seconds to start. We then check that all the
cpus we tried to start are in the online cpu mask.

Also updated the comments in thread.h to be a bit more helpful.

Test: Ran on sherlock with old, incorrect, bootloader data.
Issue: ZX-4236
Change-Id: I535e8e51868bf486f1d7cbca3f7a4695893aab29