[fidlcat] implements zx_channel_read_etc display
create type_decoder.cc/type_decoder.h
move Colors and ErrorName into these files
Add ObjTypeName and RightsName

Now the message decoding expect an array of zx_handle_info_t
instead of an array of zx_handle_t. That means that the message
decoder can display the handle type and rights.

For all the cases where we have an array of zx_handle_t (almost
everywhere), the array is converted to an array of zx_handle_info_t
with a type ZX_OBJ_TYPE_NONE and no rights.

Later, we will be able to feed this array with the right type and rights
when we have been able to infer them.

Handles are now displayed in hexdecimal with eventually the type and
rights.

The decoding of zx_channel_read_etc uses zx_channel_info_t and,
therefore, is able to display the handles' types and rights.

Add some tests for zx_channel_read_etc and modify the existing tests
to use the new handle display.

Change-Id: I755a0655ef24a5955407e4bb6b0012c735c234e5