[media][drm] Split Content Decryption API
The existing ContentDecryptionModule APIs provided additional
capabilities that a typical client should not need access to. In
particular, it gives access to provider specific provisioning and
storage management, even if the client only needs access to the license
management. This CL includes several changes to help divide those
capabilities:

  * Move provisioning APIs to a separate protocol.
  * Remove storage management APIs (these will be added back in a
    different form later).
  * Create new service hubs for each DRM system that can vend the
    ContentDecryptionModule and Provisioner APIs.

This CL also makes some additional changes that could potentially be in
a different CL, but they aide in understanding the changes:
  * Update documentation to new markdown style.
  * Add error responses for APIs that could fail and client could take
    corrective action for.

MTWN-312 #done

Change-Id: Ic6778a27464e146862dc448bcaaff48ccbba880f