[docs] Fix remaining doc checker errors
Some doc checker errors were introduced while my previous CL was waiting
in review, which block adding doc-checker to CQ.

With this change, the only blocker will be the waiting for the next
execution of the doc-roller builder to fix some auto-generated docs. The
roller runs every day as a cron job.

Bug: INTK-932
Change-Id: I687e87784c02c7d4baf41f973802fab18eaac3b7