[appmgr] use the resolver provided directory
In the near future, the path /pkgfs/packages that is used by PackageLoader
will no longer contain ephemerally fetched software. In order for ephemeral
software to be correctly serviced, it is thus a requirement that sysmgr
begins to respect the package directory handle provided by the resolver.

This change also has the added side effect of reducing the IPC and resource
overhead associated with opening packages, speeding up component launch.

There is also a small incidental change that removes no-longer used support
for non-URL component resolution.

Test: package_updating_loader_test for the package_updating_loader changes
Test: system still boots and operates for the PackageLoader, which has no
      base test infrastructure to extend for coverage. The PackageLoader
      should be no observable behavior change.
Change-Id: I36af45e8fe7f61d483dad102069dd76598badbaf