[media_codec] closure_queue source_set
This CL splits closure_queue (which was previously a part of
the codec_impl source_set) into a separate source_set, so
that it can be included in additional build targets without
requiring them to pull in all of codec_impl.

In subsequent CL, virtual_audio_driver will use
closure_queue to manage issues with asynchronous shutdown
that are detailed in MTWN-391 and other bugs.

MTWN-391 #comment

Test: no behavior change

Change-Id: I14d4114cc5f6ca92c3eecca3f1a6e14b33db8c5d