[ftdi-i2c] Add blocking Reads and Writes
Improve the performance of the FTDI Mpsse
by having the reads and writes block on the
FTDI's serial READABLE and WRITABLE.
The USB stack takes upwards of 15ms to return
so blocking is much preferred to spin-waiting.

This improves the CPU performance of constant
Read/Writes from 100% CPU usage down to 3%
CPU usage. (As measured when running
dotmatrix_display.cmx space_invaders)

TEST: runtests -t ftdi-i2c-test-test

Change-Id: I73735b2e5c350ae18e129462388d6c4f93a0bdfa