[wlan][iwlwifi] Remove page order field
Page orders aren't relevant in Fuchsia, we only care about size (in
bytes). We also don't need to track the size of buffers that closely
since each io-buffer knows its own size. We do need to translate between
the AMSDU parameter and an actual size, but only when allocating buffers
(which will only happen once). So we can safely remove this field.

BUG= WLAN-895
TEST=No functionality change. Removed dead code.

Change-Id: Icd8830830b84439113f7d9c158c880edeb3e3e0e