[scenic] View Bounds Wireframe Display
Allow users to visualize their view bounds by turning
on wireframe debugging via a scenic command.

Turning on wireframe view bounds for a view will display
that view's bounds in addition to the bounds for its
immediately embedded view holders.

This is the scenic side of the implementation only - the
functionality still needs to be exposed in Flutter, which
will come in a separate CL.

Testing - New unit test in view_clipping_unittest.cc and new
pixel test in scenic_pixel_tests.cc

SCN-1351 #comment

Change-Id: I3e1f32c75bcf6c8b95ff892415eb19ff40618d1f