[netstack3][ndp] Parameter configurations for Router Advertisements
BUG: NET-2169 #inprogress

This change just introduces the `NdpRouterConfigurations` struct to hold the values
that will be sent in Router Advertisements by a router. No functionality changed in
this change.

Tests: Unittests for `NdpRouterConfigurations` getters and setters to make sure
values are only updated if they do not violate the RFC (RFC 4861 section 6.2.1).

Change-Id: Idd3e96f32d5b3cd4b04dc9e8cf5be84329e2274f