[layout] Add a 'testing/' directory usable by tests in subareas
This adds rules for a 'testing/' directory within an area that can
contain reusable test utilities and libraries for the use of tests
within that area or subareas of that area. This is similar to the
'lib/' directory in that it can be used by subareas. It is different in
that it can only be used by targets marked 'testonly=true' in GN and
that it can contain other types of content in addition to libraries,
such as data or full component implementations.

This adds new text to the documentation and new logic (plus tests) to
the dependency checker script. This also cleans up a few exceptions
that are now allowed.

This rule does yet not work properly for rust_binary targets that
express test-only dependencies with the test_deps since that collapses
to a single cargo invocation that builds both the production and test
binaries. The checker will have to be taught about this pattern.

Change-Id: I7c9b0989c91e3115253ca6769fa7060be01384c5