[ISP] Combining OnPrimary/SecondaryFrameWritten
We had seperate functions for OnPrimaryFrameWritten and
OnSecondaryFrameWritten, because there are individual interrupts
for each part of the frame.  However, we don't use those
interrupts, so I am removing the two functions in favor of
the function they call.  This simplifies many call sites.

Testing:  This code is tested with a unit test. See
src/camera/TESTING.md for instructions on how to run the test.

Change-Id: Ie5d17f1ccd41d97270bb09d8289d68b331c2791f