[zircon][dev][goldfish] Multi-display support
This makes it possible to specify the display devices exposed
using a kernel command line variable. It makes it possible to
also specify refresh rate for each display. This refresh rate
is simulated inside the guest and will not reflect the real
refresh rate on the host.

This also adds a mode to display-test where largest display
will be mirrored. This is used when --simple flag is passed to
display-test.

Test: fx aemu -c "driver.goldfish.displays=800x600@75,640x480-800+0"
Test: display-test --simple
Change-Id: I05e792f6b3d28ada878e2a9ee6740b4e27ae7cdf