[dev][serial] Fix overflow of socket buffers
When the output serial buffers are not all written, the serial driver
saves time by reducing the size of the input buffer from the client and
marks the offset where serial writes should resume.

The space left in this buffer was miscalculated when reading from the
client, causing ZX_ERR_INVALID_ARGS when reading from the socket.

This can theoretically also happen if the socket is not written fast
enough but is unlikely because of the lack of socket backpressure.

BT-2247 #done

Test: Added SerialDeviceTest.SocketLargeWrite, revised other tests.

Change-Id: I96a1b92dc9fc9b7e63de76e54c9969dd1336b8c5