[codec][drm] Add decryptor adapter
In order to enable use of CodecImpl as a decryptor StreamProcessor, we
need a CodecAdapter implementation that can perform decryption rather
than encoding or decoding. Decryptors are reliant on the DRM key systems
that implement their functionality, but their handling of the
CodecAdapter protocol is fairly uniform. This is unlike encoders and
decoders as their buffer constraints can vary quite a bit. This CL adds
a DecryptorAdapter, which provides an implementation of the CodecAdapter
that handles the common functionality and adds a new interface specific
to decryption. DRM key systems will still need to provide an
implementation of the DecryptorAdapter that is compatible with their
system.

Change-Id: Ica6f7e98036cd2cd4d4a1e1d6cee332fc5b809d4