[modular][basemgr_launcher] Exit when basemgr is initialized
We can exit basemgr_launcher when basemgr has initialized. This can be
determined when basemgr mounts its out directory and triggers an
OnDirectoryReady event.

TEST=
basemgr_launcher relies on stdin to receive a modular configuration.
Because of this, we can't launch basemgr_launcher as a component in a
regular gtest because components don't have access to stdin.

This change was tested manually:
  `fx shell killall basemgr.cmx`
  `cat config.json | fx shell basemgr_launcher`
  `fx shell ps | grep basemgr`

Observed
  1. `cat config.json | fx shell basemgr_launcher` no long hangs
  2. The result of `fx shell ps | grep basemgr` was only one instance of
     basemgr.cmx and no instances of basemgr_launcher

MF-359 #done

Change-Id: Id8141e3f8aef65161a2f641d6775243b1c34041b