[fidlgen] Add types to represent identifiers with member names
In https://fuchsia-review.googlesource.com/c/fuchsia/+/299869,
flat::Names can now hold the name of the object member being
referenced. This needs to be accessible to code generators to
produce an appropriate output.

The goal of this CL is to support a gradual transition to this
format, given that the dart code generator is in a separate
repository.

Test: fx run-host-tests fidlgen_types_test

Change-Id: I6cc03052f41503b5bc4765232e5ed4e88da6089b