[zxtest] Print int ID when ZX_ASSERT_OK() fails.
When ZX_ASSERT_OK() fails it uses zx_status_get_string to print what the
failed status was. If that status isn't in the internal switch() of
zx_status_get_string the test just returns "(UNKNOWN)" and you have to
add extra logging to find out what the returned status was.

This adds the integer id to the printed value of the status.

Test: Ran in a failing unit test I have now and it prints the status
code as expected.

Change-Id: I8453b84869f62d19aa613f9b8cd4e1a3cbeeb192