[wlan][iwlwifi] port mvm_up and mvm_start calls.
Most of code had been merged in the previous CLs. This CL is the final
one to clean up and finish up rest of code.

Note that a lot of portions in this change are just style changes
automatically done by 'fx format'. Sorry it is hard to remove them
from this CL.

BUG=WLAN-895
TEST=Bring-up code. Doesn't impact the production code.

Change-Id: I4bfe5cce71a150052ed365356a8a36ea31db9f81