[devcoord] Add unbind_self to ScheduleRemove.
This will be used for implementing fuchsia.device.ControllerUnbind.

I also renamed fuchsia.device.Controller.Unbind to ScheduleUnbind.
This is so callers don't expect it to be synchronous, which will be
the case once we migrate to devcoord driven unbind.

I will update the tests that will be affected in additional CLs.

Change-Id: Idc222aad8b889f2e283b7e8e4b785e734d307059