[a11y] Implements LogEngine.
This change implements LogEngine, a simple tool to connect to the Tts manager as
an Engine, and listen for incoming requests.

Unlike a real Speech Engine which produces audio output, this one only prints
logs.

***
In addition, this change also implements the method Speak() under
tts_manager_impl.h, which was empty.

    Test: fx run-test a11y_lib_tests

MI4-2525 #comment

Change-Id: Ifad7469844380e426337d874608a51cb46650b0f