[sdk][tools] Strip host tools included in SDK
This is a roll-forward of the roll-back found in:

https://fuchsia.googlesource.com/build/+/ae4404dc3f7bce8f71c8d803d2a96f6c9faecd35%5E%21/#F0

The problem with that CL was that it stripped symbols that
are explicitly needed by Flutter testing on Mac builds.
This CL does not strip global symbols, which seems to make
it pass tests.

There may be alternative solutions on the Flutter side
(perhaps __attribute__((used)) works?).

DX-555 #comment

Change-Id: Idaf02c96c75646fff157d1751059e135c62024ae