[rtc] RTC drivers apply build provided backstop
On systems which have an RTC, `devmgr.epoch` was failing to have the desired
effect, as the RTC driver would later start up and take ownership over the
clock, which only applied validation based on it's constant value backstop.

`devmgr.epoch` is removed, and a new command line variable `clock.backstop`
is introduced. Both devcoordinator and librtc read this value in order to
apply UTC backstop time.

Bug: FLK-445
Test: rtc-lib-test
Test: set Acer clock to Jan, applied patch, observed applied RTC backstop
Change-Id: I9b8209b1cf984c7388dd190347f3b1cc39c3dcd8