[scenic] Working to remove Session* from Resource.
The first step is to remove Resource::error_reporter(), which obtains
the returned ErrorReporter from the Session.  The large majority of
these uses occur during command-application, and are addressed by
adding the reporter to methods on Resource subclasses.  Since this
can't be done everywhere, for safety we now use shared_ptr instead
of raw pointers to refer to ErrorReporters.

Similar work needs to be done to remove Resource::event_reporter();
this is deferred to a later CL (exception: we also replace raw ptrs
with shared_ptr for EventReporters).

The other significant change is the introduction of ImagePipeUpdater.
This allows ImagePipe to schedule frame-rendering without needing
to go through the Session*.

This CL introduces no behavior changes.

SCN-1219 #comment
SCN-1504 #comment

Change-Id: I5ab29cc07c9230a656d019e9b370a873f55b9daf