[sl4f_client] Support ssh via ssh-agent
Also throws an exception if there's no agent socket defined and no
private key was passed.

This is useful for environments in which ssh-agent manages the private
keys, like when using a Security Key's ssh key.

TESTED=ran unit tests, ran test on a device configured to allow an ssh-agent-managed public key

Change-Id: I332a4f0e8f6cdf2f7a9adb3aaab10b7ad3f340b7