[audio_core] Tighten audio fidelity testing
This CL eliminates a workaround in the audio fidelity tests that
previously added complexity to the buffer-size calculations. Now,
instead of making source buffers 1-2 frames longer than they needed to
be, the fidelity tests rely on the fact that they have created source
buffers that are perfectly periodic -- and thus when reading them they
can "wrap around" to the beginning of the source buffer when more data
is needed (as is the case when working with resamplers that have a
positive filter width, for example).

Related to this, the resampling ratio used to validate "extreme"
downsampling (~4:1) is now 191999:48000 instead of 191998:48000.
Because this is a stricter (more difficult) resampling ratio, fidelity
test thresholds for this ratio have been updated accordingly.

This test-only change prepares the way for adding the system low-pass
filter and additional resamplers.

Test: no product change, tightens existing audio_mixer_tests cases

Change-Id: I4d9a0eba152223df5ec0277b728ec1d91fe0162d