[warnings] Fix some unused variable warnings and assert side-effect issues.
Fix a couple of places where we had code patterns which said something
like...

  bool result = DoAThing();
  assert(result);

Which becomes a warning when asserts are disabled.
Also fix a couple of places where there was more serious error of the
form...

  assert(DoImportantStuffWithSideEffects());

Functional code should _never_ be placed inside of an assert as the
asserts disappear in the release code.  It might be a good idea to
audit all of the code in this module to make sure that this pattern is
not repeated in other places.  One which was discovered because of an
unused variable warning, and another which would not have been.

Change-Id: I2e0c80b95ef59a8b77adbcd3480d50df8be23d1b