[audio_core] Out-of-band Rejection tests
This CL creates a new category of audio fidelity test to assess the
system's Out-of-Band Rejection. When down-sampling, an anti-aliasing
filter should be used before decimation, to eliminate any high-
frequencies.  Because the system does not yet use an anti-aliasing
filter, the initial fidelity test thresholds reflect this state.
However, when the filter is added (in future CL), these thresholds
will be tightened accordingly.

In a related change, now that these new tests assess our out-of-band
fidelity, this CL limits the frequency response tests to in-band
frequencies only.

And finally, in another related change, an additional out-of-band
frequency is added to our set -- 95.999 kHz, just under the
Nyquist limit for our Down0 (191999:48000) resampling ratio, and
should be entirely eliminated when resampling to 48000 Hz.

Test: test-only CL, creates 3 new test cases for each resampler.

Change-Id: I6d7d892d503eea129602cf9d3b4362a751ae1778