[fidlc] Discover library dependencies used via request/response parameters
Previously, we added the containing library of composed protocols to the
list of dependencies of a library during JSON IR generation, to support
fidlgen. This CL handles adding the containing library of request and
response types of the composed protocols too.

TEST:
$FUCHSIA_DIR/out/default.zircon/host-x64-linux-clang/obj/system/utest/fidl-compiler/fidl-compiler-test.debug

Added a test where library 'bottom' declares a type, which is used as a
method parameter in library 'middle', whose protocol is composed by
library 'top'. 'top' should depend on 'bottom'

Fixes the issue in
https://fuchsia-review.googlesource.com/c/fuchsia/+/304392

Change-Id: I05f1fda73a1c36721b2f695a32ec6bf58b693d1e