[netstack3] Run benchmarks as tests in Fuchsia builds
Previously, we compiled but didn't run benchmarks when the Cargo
`benchmark` feature is disabled. As a result, some bugs were still
getting through. With this change, we put in place a framework for
running benchmarks as normal tests when the `benchmark` feature is
disabled.

We don't yet use this - the generated tests are marked `#[ignore]`.
We follow up in the next commit with a fix to the broken benchmarks
and we remove the `#[ignore]` attribute.

Change-Id: I62d86b2fa089d2aea7d0ad03ae4aefb4c5a222fb