[netdump] Eliminate non-public library dependency
CONN-158 aims to bring integration tests to netdump by moving it out
of Zircon. The first step in doing this is to make netdump only depend
on Zircon libraries available through `//zircon/public/`.

Fortunately netdump only depends on one library that is not public:
`zircon/lib/inet6.h`. This dependency was removed by replacing it with
`netinet/ip6.h` and refactoring type and variable names to equivalents.
This also promotes uniformity as IPv6 was the only exception in that
netdump uses the header structs for all other protocols from `netinet`.

During refactoring of `filter.cc`, a redundant function was also
discovered and removed.

Bug: CONN-158
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.
This CL is not expected to change observable behavior.

Change-Id: I615e5f3b6bf2d993372f1b53d3ac11e528877ea4